% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp-ratio.R
\encoding{UTF-8}
\name{asp_ratio}
\alias{asp_ratio}
\title{Compute the aspect ratio of spatial input}
\usage{
asp_ratio(x)
}
\arguments{
\item{x}{A \code{\link[terra:rast]{SpatRaster}}, an \code{\link[sf:st_sf]{sf}} or
\code{\link[sf:st_sfc]{sfc}} object or a numeric vector of length 4 with coordinates
\code{c(xmin, ymin, xmax, ymax)}, as created by \code{\link[sf:st_bbox]{sf::st_bbox()}}.}
}
\value{
A \code{\link[base:numeric]{numeric}} scalar giving the aspect ratio.
}
\description{
Compute the aspect ratio as width divided by height or columns divided by
rows.
}
\examples{
\donttest{
library(terra)

x <- rast(system.file("tiff/elev.tiff", package = "rasterpic"))
plot(x)
asp_ratio(x)
}
}
\seealso{
Other image geotagging tools:
\code{\link[=rasterpic_img]{rasterpic_img()}}
}
\concept{geotagging}
\keyword{internal}
