% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterpic-package.R
\docType{package}
\encoding{UTF-8}
\name{rasterpic-package}
\alias{rasterpic}
\alias{rasterpic-package}
\title{rasterpic: Convert Digital Images into Geotagged 'SpatRaster' Objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Geotag digital images and return 'SpatRaster' objects, as defined by the 'terra' package, using coordinates from supported spatial input classes. Supported inputs include numeric coordinate vectors and objects from the 'sf', 'terra' and 'stars' packages. The main function is an S3 generic, so other packages can provide methods for additional spatial classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/rasterpic/}
  \item \url{https://github.com/dieghernan/rasterpic}
  \item Report bugs at \url{https://github.com/dieghernan/rasterpic/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

}
\keyword{internal}
