% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp_ratio.R
\encoding{UTF-8}
\name{asp_ratio}
\alias{asp_ratio}
\title{Compute aspect ratio for spatial input}
\usage{
asp_ratio(x)
}
\arguments{
\item{x}{A \code{SpatRaster} object, an \code{sf}/\code{sfc} object or a numeric vector of
length 4 with coordinates \code{c(xmin, ymin, xmax, ymax)}, as created by
\code{\link[sf:st_bbox]{sf::st_bbox()}}.}
}
\value{
A numeric scalar giving the aspect ratio.
}
\description{
Computes the aspect ratio as width divided by height, or columns divided by
rows.
}
\examples{
\donttest{
library(terra)

x <- rast(system.file("tiff/elev.tiff", package = "rasterpic"))
plot(x)
asp_ratio(x)
}
}
\keyword{internal}
