% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterpic-package.R
\docType{package}
\name{rasterpic-package}
\alias{rasterpic}
\alias{rasterpic-package}
\title{rasterpic: Convert Digital Images into 'SpatRaster' Objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create 'SpatRaster' objects, as defined by the 'terra' package, from digital images using a spatial object as a geographic reference. Supported inputs include objects from the 'sf', 'terra' and 'stars' packages. The main function is an S3 generic, so other packages can provide methods for additional spatial classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/rasterpic/}
  \item \url{https://github.com/dieghernan/rasterpic}
  \item Report bugs at \url{https://github.com/dieghernan/rasterpic/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

}
\keyword{internal}
