% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterpic-package.R
\docType{package}
\name{rasterpic-package}
\alias{rasterpic}
\alias{rasterpic-package}
\title{rasterpic: Convert Digital Images into 'SpatRaster' Objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate 'SpatRaster' objects, as defined by the 'terra' package, from digital images, using a specified spatial object as a geographical reference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/rasterpic/}
  \item \url{https://github.com/dieghernan/rasterpic}
  \item Report bugs at \url{https://github.com/dieghernan/rasterpic/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
