% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterpic-package.R
\docType{package}
\name{rasterpic-package}
\alias{rasterpic}
\alias{rasterpic-package}
\title{rasterpic: Create a Spatial Raster from Plain Images}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Creates a spatial raster, as the ones provided by 'terra', from regular pictures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/rasterpic/}
  \item \url{https://github.com/dieghernan/rasterpic}
  \item Report bugs at \url{https://github.com/dieghernan/rasterpic/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
